move to legacy: logs
